refactor: [P2-177] force renaming summary folder
